requirements.txt
crewai
langchain
openai
yfinance
pandas
tavily-python
python-dotenv
ta
streamlit
plotly
langchain-community
langchain-openai